Run rest-api and react-viewer tests on GHA, reorganize test tree
