+++ .github/workflows/rest-api-CI.yaml
@@ -1,0 +1,57 @@
+name: rest-api_CI
+
+on:
+  push:
+    branches: ['**']
+    paths:
+      - 'wrappers/rest-api/**'
+      - '.github/workflows/rest-api-CI.yaml'
+  pull_request:
+    branches: ['**']
+    paths:
+      - 'wrappers/rest-api/**'
+      - '.github/workflows/rest-api-CI.yaml'
+
+permissions:
+  contents: read
+
+jobs:
+
+  #--------------------------------------------------------------------------------
+  React_Viewer_Tests:  # react-viewer Vitest suite (MSW-mocked, no build, no devices)
+    runs-on: ubuntu-22.04
+    timeout-minutes: 20
+    defaults:
+      run:
+        working-directory: wrappers/rest-api/tools/react-viewer
+    steps:
+    - uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd #v6.0.2
+      with:
+        persist-credentials: false
+
+    - name: Install
+      run: npm ci
+
+    - name: Run Vitest
+      run: npm test
+
+
+  #--------------------------------------------------------------------------------
+  Rest_API_Tests:  # non-live rest-api pytest (mocked devices, no build, no devices)
+    runs-on: ubuntu-22.04
+    timeout-minutes: 20
+    steps:
+    - uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd #v6.0.2
+      with:
+        persist-credentials: false
+
+    - name: Install
+      run: |
+        pip3 install -r wrappers/rest-api/requirements.txt -r unit-tests/wrappers/rest-api/requirements.txt
+        # app.services.rs_manager imports pyrealsense2. It is left out of requirements.txt
+        # so LibCI keeps using the locally-built module; GHA does not build, so pull the wheel.
+        pip3 install pyrealsense2==2.58.3.10794
+
+    - name: Run non-live tests
+      working-directory: wrappers/rest-api
+      run: python3 -m pytest tests/ -v --ignore=tests/live/
