+++ wrappers/rest-api/tests/api/__init__.py
@@ -1,0 +1,2 @@
+# License: Apache 2.0. See LICENSE file in root directory.
+# Copyright(c) 2026 RealSense, Inc. All Rights Reserved.
