+++ wrappers/rest-api/tests/manager/test_frame_metadata.py
